Normalized common forms and can fix in batch only
