+++ po-processing/src/app/app.component.html
@@ -14,36 +14,7 @@
   <!-- Warehouse Inventory Input -->
   <h2>1. Warehouse Inventory Data</h2>
   <div class="form-group row">
-    <div class="col-4 border-right">
-      <b>Upload Warehouse Excel</b>
-      <div class="input-group">
-        <div class="custom-file">
-          <!-- placeholder change detector to trigger file name rendering -->
-          <input
-            class="custom-file-input"
-            type="file"
-            id="file"
-            accept=".xls,.xlsx,application/vnd.ms-excel"
-            (change)="($event)"
-            #warehouseUpload
-          />
-          <label class="custom-file-label" for="file">
-            {{ (warehouseUpload?.files)[0]?.name || 'Choose file...' }}
-          </label>
-        </div>
-      </div>
-      <!-- TODO: parse upload xlsx -->
-      <button
-        *ngIf="(warehouseUpload?.files)[0]"
-        class="btn btn-primary mt-2"
-        (click)="
-          uploadWarehouseExcelFile(warehouseUpload); warehouseUpload.value = ''
-        "
-        disabled
-      >
-        Use This File (not available yet)
-      </button>
-    </div>
+    <div class="col-4 border-right">(Upload Feature coming soon)</div>
 
     <div class="col-8">
       <app-common-form
@@ -83,7 +54,7 @@
   <!-- Purchase Order Set Input -->
   <h2>2. Fill all POs info</h2>
   <div class="row">
-    <div class="col-4">(Upload Feature coming soon)</div>
+    <div class="col-4 border-right">(Upload Feature coming soon)</div>
 
     <div class="col-8">
       <app-common-form
@@ -110,7 +81,7 @@
             <thead class="thead-dark">
               <tr>
                 <th *ngIf="po?.shipTo">
-                  Ship to:<br />
+                  <!-- Ship to:<br />
                   {{ po.shipTo }}
                   <div class="btn-group float-right">
                     <button
@@ -123,7 +94,7 @@
                     <button class="btn btn-danger" (click)="removePo(i)">
                       Remove
                     </button>
-                  </div>
+                  </div> -->
                 </th>
               </tr>
             </thead>
